requirements.txt
grpcio
grpcio-tools
protobuf

numpy
matplotlib
Pillow
tensorflow
opencv-python

